Reorganize tasks
- Modularize tasks: new task should extend abstract class Task
- Use torchvision ResNet
- Package now has auto_train.tasks which contains the tasks (only 1 for now)
