+++ environment/src/tasks/__init__.py
@@ -1,0 +1,1 @@
+from .resnet_task import ResNetTask
